requirements.txt
aiogram>=3.4.1
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
